modify service

+++ src/main/java/com/canbro/coffee/pojo/entity/Prod.java
@@ -60,12 +60,12 @@ public class Prod implements Serializable {
     /**
      * 创建时间
      */
-    private Integer createTime;
+    private Long createTime;
 
     /**
      * 修改时间
      */
-    private Integer modifyTime;
+    private Long modifyTime;
 
     /**
      * 备注
