+++ src/main/java/com/canbro/coffee/pojo/entity/ProdAttributeKey.java
@@ -40,10 +40,10 @@ public class ProdAttributeKey implements Serializable {
     /**
      * 创建时间
      */
-    private Integer createTime;
+    private Long createTime;
 
     /**
      * 修改时间
      */
-    private Integer modifyTime;
+    private Long modifyTime;
 }
