+++ src/main/java/com/canbro/coffee/pojo/entity/Wallet.java
@@ -66,7 +66,7 @@ public class Wallet implements Serializable {
     /**
      * 创建时间
      */
-    private Integer createTime;
+    private Long createTime;
 
     /**
      * 修改者
@@ -76,7 +76,7 @@ public class Wallet implements Serializable {
     /**
      * 修改时间
      */
-    private Integer modifyTime;
+    private Long modifyTime;
 
     /**
      * 备注
